// convex/_generated/api.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
// This file is auto-generated by Convex.
// Run `npx convex dev` to regenerate this file.

// Stub API - these will be properly generated when you run `npx convex dev`
export const api: any = {
  projects: {
    list: "projects:list",
    get: "projects:get",
    create: "projects:create",
    update: "projects:update",
    remove: "projects:remove",
  },
  documents: {
    list: "documents:list",
    get: "documents:get",
    create: "documents:create",
    update: "documents:update",
    remove: "documents:remove",
  },
  revisions: {
    list: "revisions:list",
    get: "revisions:get",
    create: "revisions:create",
    restore: "revisions:restore",
  },
  personas: {
    list: "personas:list",
    get: "personas:get",
    getDefault: "personas:getDefault",
    create: "personas:create",
    update: "personas:update",
    remove: "personas:remove",
  },
  spending: {
    getToday: "spending:getToday",
    getRange: "spending:getRange",
    record: "spending:record",
  },
  userSettings: {
    get: "userSettings:get",
    upsert: "userSettings:upsert",
    clearApiKey: "userSettings:clearApiKey",
  },
};

export const internal: any = {};